services/agent_service/pyproject.toml
[project]
name = "field-service-document-intelligence"
version = "0.1.0"
description = "AI-powered field service assistant for RiteCare"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0",
    "google-genai>=1.0",
    "mcp>=1.0",
    "motor>=3.5",
    "pymongo>=4.8",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "slack-bolt>=1.20",
    "pypdf",
    "sentence-transformers>=3.0",
    "tenacity>=8.0",
    "pybreaker>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
    "ruff>=0.6",
    "mypy>=1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["shared", "db", "agent", "ritecare_tools", "agent_api"]